make the page into a map of rows, to include multiple water drop
