+++ public/manifest.json
@@ -9,7 +9,7 @@
     "name":"Ricochet",
     "url":"https://www.ricochet.exchange/"
   },
-  "Access-Control-Allow-Origin": "/*",
+  "Access-Control-Allow-Origin": "*",
   "Access-Control-Allow-Methods": "GET",
   "Access-Control-Allow-Headers": "X-Requested-With, content-type, Authorization",
   "icons":[],
